# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vector DB HTTP API
  version: 0.1.0
  description: REST interface for the on-disk Vector DB.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            text/plain:
              schema:
                type: string
  /collections:
    get:
      summary: List collections
      responses:
        '200':
          description: Collections in the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionsResponse'
    post:
      summary: Create collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreateRequest'
      responses:
        '201':
          description: Collection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /collections/{name}:
    parameters:
      - $ref: '#/components/parameters/CollectionName'
    get:
      summary: Get collection details
      responses:
        '200':
          description: Collection metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update collection settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdateRequest'
      responses:
        '200':
          description: Collection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete collection
      responses:
        '204':
          description: Collection deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /collections/{name}/rename:
    parameters:
      - $ref: '#/components/parameters/CollectionName'
    post:
      summary: Rename a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionRenameRequest'
      responses:
        '200':
          description: Collection renamed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /collections/{name}/vectors:
    parameters:
      - $ref: '#/components/parameters/CollectionName'
    post:
      summary: Insert embedding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '201':
          description: Embedding inserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /collections/{name}/vectors/{id}:
    parameters:
      - $ref: '#/components/parameters/CollectionName'
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: uint64
    get:
      summary: Fetch an embedding
      responses:
        '200':
          description: Embedding retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Replace an embedding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '200':
          description: Embedding updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete an embedding
      responses:
        '204':
          description: Embedding deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /collections/{name}/search:
    parameters:
      - $ref: '#/components/parameters/CollectionName'
    post:
      summary: Search for nearest neighbours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    CollectionName:
      name: name
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CollectionCreateRequest:
      type: object
      required: [name, dimension]
      properties:
        name:
          type: string
        dimension:
          type: integer
          minimum: 1
    CollectionResponse:
      type: object
      required: [name, dimension, vectors]
      properties:
        name:
          type: string
        dimension:
          type: integer
        vectors:
          type: integer
          minimum: 0
    CollectionRenameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    CollectionUpdateRequest:
      type: object
      required: [dimension]
      properties:
        dimension:
          type: integer
          minimum: 1
    CollectionsResponse:
      type: object
      required: [collections]
      properties:
        collections:
          type: array
          items:
            $ref: '#/components/schemas/CollectionResponse'
    EmbeddingRequest:
      type: object
      required: [id, vector]
      properties:
        id:
          type: integer
          format: uint64
        vector:
          type: array
          items:
            type: number
            format: float
        metadata:
          type: object
          additionalProperties:
            type: string
    EmbeddingResponse:
      allOf:
        - $ref: '#/components/schemas/EmbeddingRequest'
      required: [metadata]
      properties:
        metadata:
          type: object
          additionalProperties:
            type: string
    SearchRequest:
      type: object
      required: [query, k]
      properties:
        query:
          type: array
          items:
            type: number
            format: float
        k:
          type: integer
          minimum: 1
    SearchResult:
      type: object
      required: [id, distance, embedding]
      properties:
        id:
          type: integer
          format: uint64
        distance:
          type: number
          format: float
        embedding:
          $ref: '#/components/schemas/EmbeddingResponse'
    SearchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []